% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.R
\name{add.symmetric.column}
\alias{add.symmetric.column}
\title{Attaches the symmetric column to the dataset}
\usage{
add.symmetric.column(dataset, table.mapping)
}
\arguments{
\item{dataset}{A data.frame.}

\item{table.mapping}{A data.frame.}
}
\value{
new dataset with symmetric data
}
\description{
Attaches the symmetric column to the dataset
}
\examples{

}
