% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{bias_variance_ETE_plot}
\alias{bias_variance_ETE_plot}
\title{bias_variance_ETE_plot}
\usage{
bias_variance_ETE_plot(bias_variance_ETE)
}
\arguments{
\item{bias_variance_ETE}{The output of \code{compute_bias_variance_ETE}}
}
\value{
\code{ggplot} object containing bias/variance/ETE plot
}
\description{
Create a bias/variance/ETE plot based on output of \code{compute_bias_variance_ETE}
}
