% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{bias_variance_tradeoff}
\alias{bias_variance_tradeoff}
\title{Create bias-variance tradeoff plot}
\usage{
bias_variance_tradeoff(f, sigma, n, df_values = 1:15, resamples = 100)
}
\arguments{
\item{f}{The true function}

\item{sigma}{The noise level}

\item{n}{The training sample size}

\item{df_values}{The degrees of freedom values to use}

\item{resamples}{The number of resamples to use}
}
\value{
The bias-variance tradeoff plot
}
\description{
Create bias-variance tradeoff plot
}
