% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_varianza.R
\name{media_}
\alias{media_}
\alias{var_}
\alias{stat_}
\title{Calcoli Statistici: Media e Varianza}
\usage{
media_(x, p = NULL, mnam = "\\\\mu", semp = F)

var_(x, p = NULL, mnam = "\\\\sigma^2", semp = F)

stat_(x, p = NULL, mnam = "\\\\mu", vnam = "\\\\sigma^2", semp = F)
}
\arguments{
\item{x}{Un vettore di valori numerici.}

\item{p}{Un vettore opzionale di probabilità associato a \code{x}.}

\item{mnam}{Nome simbolico da utilizzare per la media, default è \eqn{\mu}.}

\item{semp}{Un booleano; se \code{TRUE}, calcola usando la semplificazione per dati campionari.}

\item{vnam}{Nome simbolico da utilizzare per la varianza, default è \eqn{\sigma^2}.}
}
\description{
Queste funzioni forniscono dettagli sui passaggi di calcolo per la media e la varianza,
e un'analisi completa combinando entrambi i calcoli.
}
\details{
La funzione \code{media_} scrive i passaggi del calcolo della media,
mentre \code{var_} quelli della varianza. La funzione \code{stat_} chiama entrambe
per fornire un'analisi completa.
}
\examples{
x <- c(1, 2, 3, 4, 5)
p <- c(0.1, 0.2, 0.3, 0.2, 0.2)
media_(x)
var_(x)
stat_(x, p=p)

}
