% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{s2c}
\alias{s2c}
\alias{sc}
\alias{vvv}
\title{Calcoli Statistici: Varianza e Deviazione Standard}
\usage{
s2c(x)

sc(x)

vvv(x, p = NULL)
}
\arguments{
\item{x}{Un vettore di valori numerici.}

\item{p}{Un vettore opzionale di probabilità associato a \code{x}.}
}
\value{
\code{s2c} e \code{vvv} restituiscono la varianza calcolata.
\code{sc} restituisce la deviazione standard calcolata.
}
\description{
Queste funzioni forniscono il calcolo della varianza e della deviazione standard per una popolazione,
e una funzione aggiuntiva per il calcolo della varianza quando si dispone di una distribuzione di probabilità.
}
\details{
La funzione \code{s2c} calcola la varianza di una popolazione. La funzione \code{sc}
calcola la deviazione standard di una popolazione. La funzione \code{vvv} calcola la varianza
per una distribuzione di probabilità o per dati grezzi se le probabilità non sono fornite.
}
\examples{
x <- c(1, 2, 3, 4, 5)
p <- c(0.1, 0.2, 0.3, 0.2, 0.2)
s2c(x)
sc(x)
vvv(x, p)

}
