% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{chi_print}
\alias{chi_print}
\alias{chi_print_conf}
\alias{chi_test}
\alias{chi_conf}
\title{Test e Visualizzazione del Chi Quadrato}
\usage{
chi_print(dat, nome_x, nome_y, print = T)

chi_print_conf(Freq_c, Freq_0, X, Y)

chi_test(dat, alpha)

chi_conf(Freq_c, Freq_0, X, Y, alpha = 0.05)
}
\arguments{
\item{dat}{Matrice o data frame contenente i dati per `chi_print` e `chi_test`.}

\item{nome_x, }{nome_y Nomi delle variabili o delle categorie per `chi_print`.}

\item{print}{Se `TRUE`, stampa la tabella di contingenza per `chi_print`.}

\item{Freq_c, }{Freq_0 Frequenze osservate e attese per `chi_print_conf` e `chi_conf`.}

\item{X, }{Y Nomi delle variabili o delle categorie per `chi_print_conf` e `chi_conf`.}

\item{alpha}{Livello di significatività per `chi_test` e `chi_conf`.}

\item{mu1, }{mu2 Medie campionarie per i test su due campioni.}

\item{s1h, }{s2h Deviazioni standard campionarie per i test su due campioni.}

\item{n1, }{n2 Dimensioni dei campioni per i test su due campioni.}

\item{h1}{Ipotesi alternativa: può essere "\neq", ">", "<".}

\item{et}{Specifica se il test assume varianze eterogenee (`TRUE`) o omogenee (`FALSE`).}

\item{a, }{b Etichette per i gruppi, utili per l'output di `test_2c`.}

\item{um}{Unità di misura (opzionale), utile per migliorare la leggibilità dell'output di `test_2c`.}
}
\description{
Fornisce metodi per visualizzare tabelle di contingenza e per eseguire il test del chi quadrato per indipendenza e conformità.
}
\details{
- `chi_print` visualizza una tabella di contingenza arricchita con i totali per righe e colonne.
- `chi_print_conf` visualizza una tabella di confronto tra frequenze osservate e attese.
- `chi_test` esegue il test del chi quadrato per indipendenza su una tabella di contingenza.
- `chi_conf` esegue il test del chi quadrato per conformità, confrontando le frequenze osservate con quelle attese.
}
\examples{
# Creazione di una tabella di contingenza
dati <- matrix(c(10, 20, 20, 40), nrow = 2)
chi_print(dati, c("Maschio", "Femmina"), c("Fumatore", "Non Fumatore"))

# Eseguire un test del chi quadrato per indipendenza
chi_test(dati, alpha = 0.05)

# Visualizzare confronto frequenze osservate e attese
chi_print_conf(c(30, 70), c(50, 50), c("Sì", "No"), c("Fumatore", "Non Fumatore"))

# Eseguire un test del chi quadrato per conformità
chi_conf(c(30, 70), c(50, 50), c("Sì", "No"), c("Fumatore", "Non Fumatore"))

}
