% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confiance-interval.R
\name{CI.mean.base}
\alias{CI.mean.base}
\title{Interval de confiance pour la moyenne}
\usage{
CI.mean.base(n, mu, sd, empirical_sd, conf.level = 0.95)
}
\arguments{
\item{n}{la taille de l'échantillon}

\item{mu}{la moyenne de l'échantillon}

\item{sd}{l'écart-type}

\item{empirical_sd}{indique si l'écart-type est connu ou alors calculé à partir de l'échantillon}

\item{conf.level}{paramètre de confiance (Défaut: 0.95)}
}
\value{
Un vecteur contenant l'interval de confiance
}
\description{
Calcul l'interval de confiance sans echantillon.
}
\examples{
Pour  évaluer le nombre de mots d’un livre, on tire 20 pages au hasard et on y compte le nombre de mots.
On trouve, pour les 20 valeurs, une moyenne de 614 mots et un  écart-type de 26 mots.
Donner un intervalle de confiance à 95\% pour le nombre total de mots du livre sachant qu’il a 158 pages
(on admettra que l’approximation normale est satisfaisante).

CI.mean.base(158, 614, 26, empirical_sd = FALSE)
}
