% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confiance-interval.R
\name{CI.mean}
\alias{CI.mean}
\title{Interval de confiance pour la moyenne}
\usage{
CI.mean(x, sd = NULL, conf.level = 0.95)
}
\arguments{
\item{x}{le vecteur}

\item{sd}{l'écart-type (Défaut: NULL, il est calculé à partir du vecteur s'il n'est pas fourni)}

\item{conf.level}{paramètre de confiance (Défaut: 0.95)}
}
\value{
Un vecteur contenant l'interval de confiance
}
\description{
Calcul l'interval de confiance à partir d'un vecteur.
}
\examples{
On veut estimer le rendement d’un engrais pour la culture du blé.
Sur douze parcelles expérimentales, on a trouvé les rendements suivants en tonnes par hectare:

rendements <- c(7.7, 8.4, 7.8, 8.2, 7.9, 8.5, 8.4, 8.2, 7.6, 7.8, 8.4, 8.3)

CI.mean(rendements)
}
