% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confiance-interval.R
\name{CI.mean.var}
\alias{CI.mean.var}
\title{Interval de confiance pour la moyenne connaissant la variance}
\usage{
CI.mean.var(n, mu, var, conf.level = 0.95)
}
\arguments{
\item{n}{la taille de l'échantillon}

\item{mu}{la moyenne de l'échantillon}

\item{var}{la variance}

\item{conf.level}{paramètre de confiance (Défaut: 0.95)}
}
\value{
Un vecteur contenant l'interval de confiance
}
\description{
Calcul l'interval de confiance pour la moyenne à partir de la taille
d'un échantillon, d'une moyenne et de la variance
}
\examples{
Considérons une population admettant une distribution normale.
On y prélève un échantillon aléatoire simple d’effectif 16, qui donne une moyenne de 80.2.
Déterminez un intervalle de confiance à 95\% pour la moyenne :

1. si σ2 = 0.09

CI.mean.var(16, 80.2, .09)
}
