% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confiance-interval.R
\name{CI.var.sd}
\alias{CI.var.sd}
\title{Interval de confiance pour la variance}
\usage{
CI.var.sd(n, mu, sd, conf.level = 0.95)
}
\arguments{
\item{n}{la taille de l'échantillon}

\item{mu}{la moyenne de l'échantillon}

\item{sd}{l'écart-type}
}
\description{
Calcul l'interval de confiance pour la variance à partir de la taille
d'un échantillon, d'une moyenne et d'un écart-type
}
\examples{
Considérons une population admettant une distribution normale.
On y prélève un échantillon aléatoire simple d’effectif 16, qui donne une moyenne de 80.2.
Déterminez un intervalle de confiance à 95\% pour la moyenne :

2. si σ2 est inconnu, et l'écart-type de l'échantillon vaut 0.3.
Dans ce dernier cas, déterminez un intervalle de confiance pour la variance, au même niveau de confiance.

CI.var.sd(16, 80.2, .3)
}
