% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{classification_metrics}
\alias{classification_metrics}
\title{classification_metrics}
\usage{
classification_metrics(
  test_responses,
  test_predictions,
  positive_class = NULL,
  C_FP = NULL,
  C_FN = NULL
)
}
\arguments{
\item{test_responses}{test responses}

\item{test_predictions}{test predictions}

\item{positive_class}{the name of the positive class}

\item{C_FP}{cost of making a false positive}

\item{C_FN}{cost of making a false negative}
}
\value{
tibble of classification metrics
}
\description{
classification_metrics
}
