% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_tidy.R
\name{coef_tidy}
\alias{coef_tidy}
\title{Extract a tidy tibble of coefficients from model fit}
\usage{
coef_tidy(fit_object, s = NULL)
}
\arguments{
\item{fit_object}{A fit object coming from \code{lm()}, \code{glm()}, \code{glmnet()}, or \code{cv.glmnet()}}

\item{s}{An optional parameter \code{s} for \code{glmnet()} and \code{cv.glmnet()}}
}
\value{
A tibble with columns \code{feature} and \code{coefficient}
}
\description{
Extract a tidy tibble of coefficients from model fit
}
