% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\docType{data}
\name{colori_base}
\alias{colori_base}
\alias{iblue}
\alias{mblue}
\alias{ablue}
\alias{ared}
\title{Colori di base}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
iblue

mblue

ablue

ared
}
\description{
Questi colori di base sono definiti utilizzando il modello di colore RGB. 
I colori includono diverse tonalità di blu e un colore rosso.
}
\details{
I colori definiti sono:
\itemize{
  \item \code{iblue}: Una versione scurita del colore base blu.
  \item \code{mblue}: Il colore base blu.
  \item \code{ablue}: Una versione più chiara del colore base blu.
  \item \code{ared}: Il colore base rosso.
}

La funzione \code{darken} viene utilizzata per regolare la luminosità del colore base blu per creare \code{iblue}.
}
\examples{
# Visualizza i colori
plot(1,1, col = iblue, pch = 16, cex = 5,xlim=c(0,5),ylim=c(0,5))
points(2,2, col = mblue, pch = 16, cex = 5)
points(3,3, col = ablue, pch = 16, cex = 5)
points(4,4, col = ared, pch = 16, cex = 5)

}
\keyword{datasets}
