% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{compute_bias_variance_ETE}
\alias{compute_bias_variance_ETE}
\title{compute_bias_variance_ETE}
\usage{
compute_bias_variance_ETE(fitted_spline_models, f, sigma)
}
\arguments{
\item{fitted_spline_models}{Data frame outputed by \code{fit_spline_models()}}

\item{f}{The true function}

\item{sigma}{The noise standard deviation}
}
\value{
A data frame with columns \code{df}, \code{mean_sq_bias}, \code{mean_variance},
and \code{expected_test_error}
}
\description{
compute_bias_variance_ETE
}
