% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-unimore.R, R/probabilita.R
\name{two_way}
\alias{two_way}
\alias{two_way2}
\title{Convolution: Metodi Dettagliati}
\usage{
two_way(
  S_1,
  S_2,
  num1,
  num2,
  op = `+`,
  EV = T,
  vnam = "X",
  size = "\\\\normalsize "
)

two_way2(
  S_1,
  S_2,
  p1,
  p2,
  op = `+`,
  EV = T,
  vnam = "X",
  size = "\\\\normalsize "
)

two_way(
  S_1,
  S_2,
  num1,
  num2,
  op = `+`,
  EV = T,
  vnam = "X",
  size = "\\\\normalsize "
)

two_way2(
  S_1,
  S_2,
  p1,
  p2,
  op = `+`,
  EV = T,
  vnam = "X",
  size = "\\\\normalsize "
)
}
\arguments{
\item{S_1}{Insieme dei valori per la prima variabile.}

\item{S_2}{Insieme dei valori per la seconda variabile.}

\item{num1}{Frequenze o probabilità per \code{S_1} in \code{two_way}.}

\item{num2}{Frequenze o probabilità per \code{S_2} in \code{two_way}.}

\item{op}{Operatore da utilizzare nella convoluzione (default è \code{`+`}).}

\item{EV}{Se \code{TRUE}, calcola valore atteso e varianza.}

\item{vnam}{Nome variabile per l'output.}

\item{size}{Dimensione del testo per l'output LaTeX.}

\item{p1}{Probabilità per \code{S_1} in \code{two_way2}.}

\item{p2}{Probabilità per \code{S_2} in \code{two_way2}.}
}
\value{
Entrambe le funzioni stampano i dettagli della convoluzione ma non restituiscono valori.

Entrambe le funzioni stampano i dettagli della convoluzione ma non restituiscono valori.
}
\description{
Queste funzioni eseguono e stampano i passaggi di una convoluzione finita tra due insiemi di dati.
\code{two_way} utilizza le frequenze senza semplificare, mentre \code{two_way2} utilizza le frequenze calcolate.

Queste funzioni eseguono e stampano i passaggi di una convoluzione finita tra due insiemi di dati.
\code{two_way} utilizza le frequenze senza semplificare, mentre \code{two_way2} utilizza le frequenze calcolate.
}
\examples{
S_1 <- c(1, 2)
S_2 <- c(3, 4)
num1 <- c(1, 1)
num2 <- c(1, 1)
p1 <- c(0.5, 0.5)
p2 <- c(0.5, 0.5)
two_way(S_1, S_2, num1, num2)
two_way2(S_1, S_2, p1, p2)

S_1 <- c(1, 2)
S_2 <- c(3, 4)
num1 <- c(1, 1)
num2 <- c(1, 1)
p1 <- c(0.5, 0.5)
p2 <- c(0.5, 0.5)
two_way(S_1, S_2, num1, num2)
two_way2(S_1, S_2, p1, p2)

}
