% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountFive.R
\name{CountFive}
\alias{CountFive}
\title{Use two inputs to test significance level using R.}
\usage{
CountFive(x, y)
}
\arguments{
\item{x}{the first sample(vector)}

\item{y}{the second sample(vector)}
}
\value{
number of extreme points \code{n}
}
\description{
A two sample test of equal variance.
}
\examples{
\dontrun{
number <- CountFive(x,y)
}
}
