% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{cross_validate_spline}
\alias{cross_validate_spline}
\title{cross_validate_spline}
\usage{
cross_validate_spline(x, y, nfolds, df_values)
}
\arguments{
\item{x}{vector of x coordinates in training data}

\item{y}{vector of y coordinates in training data}

\item{nfolds}{number of folds for cross-validation}

\item{df_values}{vector of values of degrees of freedom to try}
}
\value{
An object containing the CV table, CV plot, df.1se and df.min
}
\description{
Run cross-validation to select the degrees of freedom
}
