% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{cv_tree}
\alias{cv_tree}
\title{Extract CV information from decision tree fit}
\usage{
cv_tree(tree_fit)
}
\arguments{
\item{tree_fit}{Tree fit object returned by rpart}
}
\value{
An object containing the CV table, CV plot, nleaves.1se and nleaves.min
}
\description{
Extract CV information from decision tree fit
}
