% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-unimore.R, R/probabilita.R
\name{bin_dis}
\alias{bin_dis}
\alias{pois_dis}
\alias{norm_int}
\title{Distribuzioni di Probabilità: Binomiale, Poisson, Normale}
\usage{
bin_dis(
  x1,
  n,
  pp,
  verso = "\\\\leq",
  comp = FALSE,
  sing = FALSE,
  x0 = 0,
  vnam = "X",
  size = "\\\\normalsize"
)

pois_dis(x1, ll, verso = "\\\\leq", sing = FALSE, vnam = "X")

norm_int(
  x1,
  x2 = NULL,
  verso = "<",
  mm,
  ss,
  vnam = "X",
  mu = "\\\\mu",
  sigma = "\\\\sigma"
)

bin_dis(
  x1,
  n,
  pp,
  verso = "\\\\leq",
  comp = FALSE,
  sing = FALSE,
  x0 = 0,
  vnam = "X",
  size = "\\\\normalsize"
)

pois_dis(x1, ll, verso = "\\\\leq", sing = FALSE, vnam = "X")

norm_int(
  x1,
  x2 = NULL,
  verso = "<",
  mm,
  ss,
  vnam = "X",
  mu = "\\\\mu",
  sigma = "\\\\sigma"
)
}
\arguments{
\item{x1, x2}{Valori di interesse per il calcolo della probabilità.}

\item{n}{Numero di prove in `bin_dis`.}

\item{pp}{Probabilità di successo in `bin_dis`.}

\item{verso}{Specifica se calcolare la probabilità cumulativa a sinistra (\eqn{\leq}) o a destra (\eqn{\geq}).}

\item{comp}{Se `TRUE`, calcola la probabilità complementare in `bin_dis`.}

\item{sing}{Se `TRUE`, calcola la probabilità per un singolo valore.}

\item{x0}{Valore iniziale per il calcolo in `bin_dis`.}

\item{vnam}{Nome della variabile casuale.}

\item{size}{Dimensione del testo LaTeX.}

\item{ll}{Parametro \eqn{\lambda} per la distribuzione di Poisson in `pois_dis`.}

\item{mm}{Media della distribuzione in `norm_int`.}

\item{ss}{Varianza della distribuzione in `norm_int`.}
}
\description{
Queste funzioni calcolano e visualizzano i passaggi per le distribuzioni di probabilità binomiale (`bin_dis`), di Poisson (`pois_dis`), e normale (`norm_int`).

Queste funzioni calcolano e visualizzano i passaggi per le distribuzioni di probabilità binomiale (`bin_dis`), di Poisson (`pois_dis`), e normale (`norm_int`).
}
\details{
\itemize{
\item `bin_dis` visualizza i passaggi per il calcolo della probabilità in una distribuzione binomiale.
\item `pois_dis` mostra i passaggi per una distribuzione di Poisson.
\item `norm_int` calcola e visualizza i passaggi per probabilità in una distribuzione normale.
}

\itemize{
\item `bin_dis` visualizza i passaggi per il calcolo della probabilità in una distribuzione binomiale.
\item `pois_dis` mostra i passaggi per una distribuzione di Poisson.
\item `norm_int` calcola e visualizza i passaggi per probabilità in una distribuzione normale.
}
}
\examples{
# Distribuzione binomiale
bin_dis(x1 = 2, n = 5, pp = 0.5)

# Distribuzione di Poisson
pois_dis(x1 = 3, ll = 2.5)

# Distribuzione normale
norm_int(x1 = -1.96, x2 = 1.96, mm = 0, ss = 1)

# Distribuzione binomiale
bin_dis(x1 = 2, n = 5, pp = 0.5)

# Distribuzione di Poisson
pois_dis(x1 = 3, ll = 2.5)

# Distribuzione normale
norm_int(x1 = -1.96, x2 = 1.96, mm = 0, ss = 1)

}
