% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilita.R
\name{draw_dist}
\alias{draw_dist}
\title{Disegna una distribuzione tratteggiata}
\usage{
draw_dist(dist, z1, z2, density = 20, border = NA, col = 1, ...)
}
\arguments{
\item{dist}{Una funzione che definisce la distribuzione da tracciare.}

\item{z1}{Il limite inferiore dell'intervallo su cui tracciare la distribuzione.}

\item{z2}{Il limite superiore dell'intervallo su cui tracciare la distribuzione.}

\item{density}{La densità delle linee tratteggiate (opzionale, predefinito è 20).}

\item{border}{Il colore del bordo del poligono (opzionale, predefinito è NA).}

\item{col}{Il colore della distribuzione (opzionale, predefinito è 1).}

\item{...}{Argomenti addizionali passati alla funzione \code{polygon}.}
}
\description{
Questa funzione aggiunge una distribuzione tratteggiata a un grafico esistente, utilizzando un poligono e una curva.
}
\details{
La funzione disegna un poligono tratteggiato per rappresentare la distribuzione tra \code{z1} e \code{z2}, e aggiunge una curva della distribuzione sopra il poligono.
}
\examples{
# Definisce una distribuzione normale standard
dist <- function(x) dnorm(x, mean = 0, sd = 1)

# Disegna la distribuzione normale tra -2 e 2
plot(0, type = "n", xlim = c(-3, 3), ylim = c(0, 0.5))
draw_dist(dist, -2, 2, col = "blue", density = 30, angle = 45)

}
