% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glmnet.R
\name{extract_best_elnet}
\alias{extract_best_elnet}
\title{extract_best_elnet}
\usage{
extract_best_elnet(elnet_fit)
}
\arguments{
\item{elnet_fit}{fit object returned by cva.glmnet}
}
\value{
glmnet fit object
}
\description{
Extract best elastic net fit over alpha
}
