% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glmnet.R
\name{extract_std_coefs}
\alias{extract_std_coefs}
\title{extract_std_coefs}
\usage{
extract_std_coefs(glmnet_fit, data, lambda = NULL)
}
\arguments{
\item{glmnet_fit}{The glmnet fit object}

\item{data}{The data}

\item{lambda}{The value of lambda}
}
\value{
A tibble with columns \code{feature} and \code{coefficient}
}
\description{
extract_std_coefs
}
