% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_spline_model}
\alias{fit_spline_model}
\title{fit_spline_model}
\usage{
fit_spline_model(data, df)
}
\arguments{
\item{data}{Data with columns x and y}

\item{df}{Degrees of freedom to use}
}
\value{
Fitted spline model
}
\description{
fit_spline_model
}
