% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{fit_spline_models}
\alias{fit_spline_models}
\title{fit_spline_models}
\usage{
fit_spline_models(train_data_resamples, df_values)
}
\arguments{
\item{train_data_resamples}{Data frame with columns \code{x}, \code{y}, and optionally, \code{resample} if multiple datasets are given}

\item{df_values}{Vector of values of degrees of freedom for spline fits.}
}
\value{
Data frame with columns \code{x}, \code{y}, \code{resample}, \code{df}, \code{pred}
}
\description{
Fit natural spline models with various degrees of freedom to one or more datasets.
}
