% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{generate_spline_data}
\alias{generate_spline_data}
\title{generate_spline_data}
\usage{
generate_spline_data(f, sigma, n, resamples)
}
\arguments{
\item{f}{Function taking values on (0, 2*pi)}

\item{sigma}{Noise standard deviation}

\item{n}{Number of points to generate}

\item{resamples}{Number of resamples}
}
\value{
Data frame with all the datasets
}
\description{
Generate \code{resamples} datasets from Y = f(X)+epsilon for
X sampled on an equi-spaced grid of \code{n} samples on (0, 2*pi),
and noise standard deviation \code{sigma}.
}
