% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genera_dati.R
\name{genera_dati}
\alias{genera_dati}
\alias{vunif}
\title{Genera Dati}
\usage{
genera_dati(brk, hhh = NULL, n, nnn = NULL, rand = T)

vunif(nnn, brk)
}
\arguments{
\item{brk}{Un vettore numerico che specifica gli intervalli per la generazione dei dati.}

\item{hhh}{Un valore opzionale utilizzato per calcolare `nnn` se quest'ultimo non è fornito.}

\item{n}{Il numero totale di dati da generare.}

\item{nnn}{Un vettore opzionale che specifica il numero di dati da generare per ogni intervallo specificato in `brk`.
Se `NULL`, viene calcolato automaticamente basandosi su `hhh`.}

\item{rand}{Un booleano che, se `TRUE`, genera dati casuali; altrimenti, genera dati uniformi.}
}
\value{
Un vettore contenente i dati generati.
}
\description{
Questa funzione genera un insieme di dati basato sui parametri forniti. Può generare dati casuali o uniformi
in base all'intervallo specificato da `brk` e, opzionalmente, basato su `hhh` e `nnn`.
}
\examples{
brk <- c(1, 5, 10)
hhh <- c(2,3)
n <- 100
dati_generati <- genera_dati(brk, hhh, n)
print(dati_generati)
}
