% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_base.R
\name{h.int}
\alias{h.int}
\alias{percentile}
\alias{F_print}
\alias{histp}
\title{Funzioni per la manipolazione e visualizzazione dei dati}
\usage{
h.int(x1, x2, density = 20, ...)

percentile(p = 0.5)

F_print(x, verso = "<", x2 = 0)

histp(axes = F, ...)
}
\description{
Queste funzioni sono utilizzate per il calcolo dei percentili, la stampa di distribuzioni cumulative e la visualizzazione di istogrammi con densità percentuale.
}
\details{
Le seguenti funzioni sono incluse:

\code{h.int}: Funzione per disegnare un'area sotto la curva interpolata.

\code{percentile}: Calcola e stampa l'approssimazione di un percentile dato.

\code{F_print}: Stampa la distribuzione cumulativa in LaTeX.

\code{histp}: Disegna un istogramma con densità percentuale.
}
\examples{
brk <- c(0,1,2,5,10)
hhh <- c(2,5,1,.5)
samp <- genera_dati(brk = brk, hhh = hhh, n = 150)
ls2e(stat_base(samp = samp, brk = brk))

# Utilizzo delle funzioni

# Disegna un'area sotto la curva interpolata

# Calcola il percentile
percentile(p = 0.5)

# Stampa la distribuzione cumulativa in LaTeX
F_print(x = 5, verso = "<")

# Disegna un istogramma con densità percentuale
histp(axes = TRUE)
h.int(x1 = 2, x2 = 8, density = 20)

}
