% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hat_alpha.R
\name{hat_alphat}
\alias{hat_alphat}
\title{The significance level for CountFive test.}
\usage{
hat_alphat(x, y, m = 1000)
}
\arguments{
\item{x}{the first sample(vector).}

\item{y}{the second sample(vector).}
}
\value{
empirical Type I error rate \code{n}
}
\description{
Estimate the significance level of the CountFive test .
}
