% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-unimore.R, R/main-functions.R
\name{item}
\alias{item}
\title{Genera Punto Elenco Personalizzato}
\usage{
item(new = FALSE, num = FALSE)

item(new = FALSE, num = FALSE)
}
\arguments{
\item{i1}{Numero intero per il punto elenco.}

\item{i2}{Indice per la lettera nel punto elenco, dove 1 = a, 2 = b, ecc.}
}
\value{
Restituisce una stringa che rappresenta il punto elenco personalizzato.

Restituisce una stringa che rappresenta il punto elenco personalizzato.
}
\description{
Questa funzione genera un punto elenco personalizzato nel formato numero.lettera.

Questa funzione genera un punto elenco personalizzato nel formato numero.lettera.
}
\examples{
item(1, 1) # Restituisce "1.a"
item(2, 3) # Restituisce "2.c"

item(1, 1) # Restituisce "1.a"
item(2, 3) # Restituisce "2.c"

}
