% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_anova_blok}
\alias{kod_anova_blok}
\title{Szkielet: ANOVA - Blokowe Losowe (Kod ze zdjęcia)}
\usage{
kod_anova_blok()
}
\description{
Dokładne odwzorowanie kodu ze zdjęcia (Lista 6 Zadanie 1).
Liczy ręcznie sumy kwadratów (SST, SSA, SSB, SSE) i statystykę F.
}
