% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_anova_grecko_lacinski}
\alias{kod_anova_grecko_lacinski}
\title{Szkielet: ANOVA - Kwadrat Grecko-Łaciński (Kod ze zdjęcia)}
\usage{
kod_anova_grecko_lacinski()
}
\description{
Dokładne odwzorowanie kodu ze zdjęcia (Zadanie 3 - Stanowisko pracy).
Obsługuje cztery źródła zmienności: Wiersze, Kolumny, Litery łacińskie i Litery greckie.
}
