% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_anova_lacinski_orig}
\alias{kod_anova_lacinski_orig}
\title{Szkielet: ANOVA - Kwadrat Łaciński (Kod ze zdjęcia)}
\usage{
kod_anova_lacinski_orig()
}
\description{
Dokładne odwzorowanie kodu ze zdjęcia (Zadanie 2 - Telewizory).
Uwzględnia macierz wartości (mat2) i macierz układu literek (literki).
}
