% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_boot_korelacja}
\alias{kod_boot_korelacja}
\title{Szkielet: Bootstrap dla KORELACJI (2 zmienne)}
\usage{
kod_boot_korelacja()
}
\description{
Generuje kod do zadania z macierzą kowariancji (rmvnorm) i korelacją.
UWAGA: Wymaga pakietu mvtnorm, jeśli generujesz dane.
}
