% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_boot_param}
\alias{kod_boot_param}
\title{Szkielet: Bootstrap PARAMETRYCZNY (Poisson, Wykładniczy, Normalny)}
\usage{
kod_boot_param()
}
\description{
Gdy znamy rozkład danych. Generuje kod z przykładami dla różnych rozkładów.
}
