% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_boot_regresja_manual}
\alias{kod_boot_regresja_manual}
\title{Szkielet: Bootstrap RESZTOWY (Wersja manualna z wektorami e1, e2, e3)}
\usage{
kod_boot_regresja_manual()
}
\description{
Szkielet: Bootstrap RESZTOWY (Wersja manualna z wektorami e1, e2, e3)
}
