% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_bootstrap}
\alias{kod_bootstrap}
\title{Szkielet: Bootstrap (Przedział Ufności i Estymatory)}
\usage{
kod_bootstrap()
}
\description{
Generuje kod do obliczenia estymatora, obciążenia i przedziału ufności metodą percentyli.
}
