% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_perm_niezalezne}
\alias{kod_perm_niezalezne}
\title{Szkielet: Permutacje NIEZALEŻNE (Dwie grupy)}
\usage{
kod_perm_niezalezne()
}
\description{
Test dla dwóch różnych grup (wrzucanie do jednego worka).
}
