% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_perm_pary}
\alias{kod_perm_pary}
\title{Szkielet: Permutacje dla PAR (Zależne, np. przed/po)}
\usage{
kod_perm_pary()
}
\description{
Test permutacyjny dla dwóch pomiarów na tych samych obiektach.
Losowanie znaków +/-.
}
