% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{leaf}
\alias{leaf}
\title{Properties of 340 leaves}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 340 rows and 17 columns.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Leaf}
}
\usage{
leaf
}
\description{
A dataset containing attributes of a pile of leaves
}
\details{
The variables are (roughly) described below. More details as well as the
actual images are available at the UCI ML repository.
\enumerate{
\item Species - A number from 1 to 36 indicating which species the leaf represents
\item Specimen Number - Numbered sequentially within species
\item Eccentricity - Eccentricity of the ellipse with identical second moments to the image. This value ranges from 0 to 1.
\item Aspect Ratio - Values close to 0 indicate an elongated shape.
\item Elongation - The minimum is achieved for a circular region.
\item Solidity - It measures how well the image fits a convex shape.
\item Stochastic Convexity - This variable extends the usual notion of convexity in topological sense, using sampling to perform the calculation.
\item Isoperimetric Factor - The maximum value of 1 is reached for a circular region. Curvy intertwined con- tours yield low values.
\item Maximal Indentation Depth - How deep indentations are
\item Lobedness - This feature characterizes how lobed a leaf is.
\item Average Intensity - Average intensity is defined as the mean of the intensity image
\item Average Contrast - Average contrast is the the standard deviation of the intensity im- age
\item Smoothness - For a region of constant intensity, this takes the value 0 and approaches 1 as regions exhibit larger disparities in intensity values.
\item Third moment - a measure of the intensity histogram’s skewness
\item Uniformity - uniformity’s maximum value is reached when all intensity levels are equal.
\item Entropy - A measure of intensity randomness.
}
}
\keyword{datasets}
