% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_cpp}
\alias{lm_cpp}
\title{Fitting Simple Linear Models}
\usage{
lm_cpp(y, x)
}
\arguments{
\item{y}{a vector of response variable.}

\item{x}{a vector of predictor variable.}
}
\value{
coefficients a named vector of coefficients.

se the standard errors.

ci the 95\% confidence intervals.

residuals the residuals, that is response minus fitted values.

fitted.values the fitted mean values.
}
\description{
'lm_cpp' is used to fit simple linear models.
}
