% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{ls2e}
\alias{ls2e}
\title{Converte una Lista in un Ambiente e Assegna gli Elementi alla Global Environment}
\usage{
ls2e(x)
}
\arguments{
\item{x}{Lista degli oggetti da convertire e assegnare nell'ambiente globale.}
}
\description{
Questa funzione prende una lista di oggetti e li assegna alla global environment, 
rendendo ogni elemento della lista direttamente accessibile come un oggetto 
indipendente nell'ambiente globale.
}
\details{
La funzione utilizza `list2env()` per convertire la lista in un ambiente, 
poi assegna questo ambiente alla global environment. Gli oggetti all'interno della 
lista diventano così direttamente accessibili nell'ambiente globale. La funzione 
opera in modo invisibile, senza produrre output diretti.
}
\examples{
my_list <- list(a = 1, b = 2, c = 3)
ls2e(my_list)
a # Dovrebbe restituire 1
b # Dovrebbe restituire 2
c # Dovrebbe restituire 3

}
