% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{marbles}
\alias{marbles}
\title{Race results from \href{https://www.youtube.com/channel/UCYJdpnjuSWVOLgGT9fIzL0g}{Jelle's Marble Runs}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128 rows and 13 columns.
}
\source{
This data is
courtesy of \href{http://www.randalolson.com/2020/05/24/a-data-driven-look-at-marble-racing/}{Randy Olson}.
See also Randy's \href{http://www.randalolson.com/2020/05/24/a-data-driven-look-at-marble-racing/}{blogpost} which
covers some additional analysis.
}
\usage{
marbles
}
\description{
Included variables (some are self explanatory):
\itemize{
\item \code{Marble_Name}
\item \code{Race_number} number of the race in Series 1
\item \code{Qualifier_Time} the marble's time in a 1-lap qualification (in seconds)
\item \code{Race_Time} final time for the race (also in seconds)
\item \code{Starting_position} based on qualifier outcome
\item \code{Race_points} allocated to the race winners
\item \code{Qualifier_Lap_Avg}
\item \code{Race_Lap_Avg}
\item \code{Site} which race track?
\item \code{Team_Name} the marbles are on different teams
\item \code{Track_length} in meters
\item \code{Host} binary, did this marble's team host the race
\item \code{Race_laps} how many laps in the race
}
}
\details{
Which marbles are most likely to win? This data contains race and qualifier results
for Season 1.
}
\keyword{datasets}
