% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{mixed_cor}
\alias{mixed_cor}
\title{Computed mixed correlation matrix}
\usage{
mixed_cor(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
Correlation matrix
}
\description{
Like a regular correlation matrix, but allowing arbitrary column types.
}
\examples{
mixed_cor(ggplot2::diamonds)
}
