% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{p}
\alias{p}
\title{Formatta Numeri con Parentesi per Valori Negativi}
\usage{
p(x, ax = 4)
}
\arguments{
\item{x}{Numero da formattare.}

\item{ax}{Numero di cifre decimali a cui arrotondare il valore di `x`. 
Il valore predefinito è 4.}
}
\value{
Restituisce una stringa che rappresenta il numero formattato,
        con parentesi intorno ai valori negativi.
}
\description{
Questa funzione formatta i numeri aggiungendo parentesi intorno ai valori negativi
e arrotonda il numero alla precisione specificata.
}
\examples{
p(-3.14159265) # Restituisce "(-3.1416)"
p(3.14159265)  # Restituisce "3.1416"

}
