% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilita.R
\name{p_ab}
\alias{p_ab}
\title{Calcolo della probabilità dell'intersezione di due eventi}
\usage{
p_ab(pa, pb, A = "A", B = "B", dig = 4)
}
\arguments{
\item{pa}{La probabilità dell'evento \eqn{A}.}

\item{pb}{La probabilità dell'evento \eqn{B}.}

\item{A}{Il nome dell'evento \eqn{A} (opzionale, predefinito è "A").}

\item{B}{Il nome dell'evento \eqn{B} (opzionale, predefinito è "B").}

\item{dig}{Il numero di cifre decimali per l'arrotondamento (opzionale, predefinito è 4).}
}
\description{
Questa funzione calcola la probabilità dell'intersezione di due eventi \eqn{A} e \eqn{B} dati le probabilità individuali \eqn{P(A)} e \eqn{P(B)}.
}
\details{
La probabilità dell'intersezione di due eventi \eqn{A} e \eqn{B} è data dalla formula:
\deqn{P(A \cap B) = P(A) \cdot P(B)}
se \eqn{A} e \eqn{B} sono indipendenti.

La funzione stampa la formula e il risultato calcolato in notazione LaTeX.
}
\examples{
p_ab(0.5, 0.3)
p_ab(0.5, 0.3, A = "Evento1", B = "Evento2", dig = 2)

}
