% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_learning_helpers.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot confusion matrix}
\usage{
plot_confusion_matrix(
  predicted_responses,
  actual_responses,
  class_names = NULL
)
}
\arguments{
\item{predicted_responses}{A vector of predicted responses for the test set}

\item{actual_responses}{A vector of actual responses for the test set}

\item{class_names}{A list of more informative class names (optional)}
}
\value{
A \code{ggplot} object plotting the confusion matrix
}
\description{
Plot confusion matrix
}
