% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glmnet.R
\name{plot_cva_glmnet}
\alias{plot_cva_glmnet}
\title{plot_cva_glmnet}
\usage{
plot_cva_glmnet(elnet_fit)
}
\arguments{
\item{elnet_fit}{fit object returned by cva.glmnet}
}
\value{
A ggplot object
}
\description{
Plot CV as a function of alpha for cva.glmnet
}
