% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilita.R
\name{plot_FdR}
\alias{plot_FdR}
\title{Traccia la funzione di distribuzione cumulativa}
\usage{
plot_FdR(sx, pp, xmin, xmax)
}
\arguments{
\item{sx}{Un vettore di valori x per i punti della distribuzione.}

\item{pp}{Un vettore di probabilità corrispondenti ai valori in \code{sx}.}

\item{xmin}{Il valore minimo di x per la trama.}

\item{xmax}{Il valore massimo di x per la trama.}
}
\description{
Questa funzione traccia la funzione di distribuzione cumulativa (F(x)) utilizzando punti e segmenti per rappresentare la distribuzione discreta data dai valori di ingresso.
}
\details{
La funzione prende i valori di ingresso \code{sx} e \code{pp} e traccia la funzione di distribuzione cumulativa utilizzando segmenti orizzontali e verticali. I valori cumulativi sono tracciati sull'asse y.
}
\examples{
sx <- c(1, 2, 3)
pp <- c(0.2, 0.5, 0.3)
xmin <- 0
xmax <- 4
plot_FdR(sx, pp, xmin, xmax)

}
