% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glmnet.R
\name{plot_glmnet}
\alias{plot_glmnet}
\title{plot_glmnet}
\usage{
plot_glmnet(glmnet_fit, data, lambda = NULL, features_to_plot = NULL)
}
\arguments{
\item{glmnet_fit}{fit object returned either by glmnet or cv.glmnet}

\item{data}{the data}

\item{lambda}{a value of the penalty parameter; if null and cv.glmnet}

\item{features_to_plot}{either number or names of features to highlight in the plot;}
}
\value{
Trace plot object
}
\description{
Create a trace plot of lasso or ridge regression, highlighting certain features.
}
