% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_learning_helpers.R
\name{plot_grayscale}
\alias{plot_grayscale}
\title{Plot a grayscale image}
\usage{
plot_grayscale(image_array, label = NULL, class_names = NULL)
}
\arguments{
\item{image_array}{An array of pixels to plot}

\item{label}{The class label, to be used as the title of the plot (optional)}

\item{class_names}{A list of more informative class names (optional)}
}
\value{
A \code{ggplot} object containing the image
}
\description{
Plot a grayscale image
}
