% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_learning_helpers.R
\name{plot_model_history}
\alias{plot_model_history}
\title{Produce a nice plot of the deep learning model training history}
\usage{
plot_model_history(history)
}
\arguments{
\item{history}{A history object, as outputed by Keras}
}
\value{
A ggplot object plotting the history
}
\description{
Produce a nice plot of the deep learning model training history
}
