% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{punt_p}
\alias{punt_p}
\title{Prepara i punteggi e le enumerazioni per i compiti}
\usage{
punt_p(start = F, nex = F, num = F)
}
\arguments{
\item{start}{logico, si inizia dall'esercizio 1?.}

\item{nex}{logico, prossimo esercizio?}
}
\value{
il numero progressivo dell'esercizio e il relativo punteggio
}
\description{
Prepara i punteggi e le enumerazioni per i compiti
}
\details{
l'oggetto \code{punti} va definito nel \code{.GlobalEnv}
}
\examples{
punti <- list(
  e1 = c(14,3,2,2),
  e2 = c(14,3,2,2),
  e3 = c(14),
  e4 = c(3,3,3,3),
  e5 = c(4,10),
  e6 = c(14,3,2,2,2)
)
punt_p(start=T)
punt_p()
punt_p()
punt_p(nex=T)

punt_p(num=T)
}
