% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R
\name{R2}
\alias{R2}
\title{Stampa e Valuta il Coefficiente di Determinazione \(r^2\) in LaTeX}
\usage{
R2(adatto = T, percent = T)
}
\arguments{
\item{adatto}{Logico: deve dire se il modello è adatto?}

\item{percent}{Logico: deve stampare la percentuale di varianza spiegata?.}
}
\value{
La funzione non ritorna un valore ma stampa direttamente l'output in formato LaTeX
        e un messaggio testuale che valuta la bontà di adattamento del modello nell'ambiente di chiamata.
}
\description{
Questa funzione calcola il coefficiente di correlazione \(r\), il suo quadrato \(r^2\),
e stampa queste statistiche in formato LaTeX. Inoltre, valuta se \(r^2\) indica un buon adattamento
del modello ai dati confrontando \(r^2\) con il valore soglia 0.75.
}
\details{
Si assume che le variabili `r`, `co`, `sx`, e `sy` (correlazione, covarianza di X e Y,
         deviazione standard di X e deviazione standard di Y, rispettivamente) siano definite e
         accessibili nello scope da cui la funzione viene chiamata. La funzione determina se
         il modello si adatta bene ai dati basandosi su se \(r^2\) è maggiore o minore di 0.75.
}
\examples{
# Calcolo e valutazione di r^2 supponendo che r, co, sx, sy siano già definiti
x <- rnorm(100)
y <- x + rnorm(100,0,.1)
ls2e(regr(x,y))
R2()

}
