% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilita.R
\name{r_norm}
\alias{r_norm}
\title{Generazione di coppie di valori con media, deviazione standard e covarianza specificate}
\usage{
r_norm(
  xbar = 0,
  ybar = 0,
  sx = 1,
  sy = 1,
  sxy = 0,
  n = 10,
  max_iter = 10000,
  tol = 1e-07,
  dist = rnorm,
  contr = FALSE,
  prec = 6,
  ...
)
}
\arguments{
\item{xbar}{Valore atteso della media di `x`. Default: `0`.}

\item{ybar}{Valore atteso della media di `y`. Default: `0`.}

\item{sx}{Deviazione standard desiderata per `x`. Default: `1`.}

\item{sy}{Deviazione standard desiderata per `y`. Default: `1`.}

\item{sxy}{Covarianza desiderata tra `x` e `y`. Default: `0`.}

\item{n}{Numero di coppie `(x, y)` da generare. Default: `10`.}

\item{max_iter}{Numero massimo di iterazioni per la convergenza della covarianza. Default: `10000`.}

\item{tol}{Tolleranza per la convergenza della covarianza. Default: `1e-7`.}

\item{dist}{Funzione di distribuzione da cui generare i valori di `x`. Default: `rnorm`.}

\item{contr}{Se `TRUE`, stampa a video le differenze tra i valori desiderati e quelli ottenuti. Default: `FALSE`.}

\item{prec}{Precisione decimale per i risultati della verifica (`contr = TRUE`). Default: `10`.}

\item{...}{Argomenti aggiuntivi da passare alla funzione di generazione `dist`.}
}
\value{
Una matrice `n x 2` in cui la prima colonna rappresenta `x` e la seconda colonna rappresenta `y`.
}
\description{
Questa funzione genera coppie di valori `(x, y)` distribuiti normalmente, 
garantendo che abbiano le medie, le deviazioni standard e la covarianza desiderate, 
attraverso un processo iterativo di correzione.
}
\examples{
set.seed(123)
r_norm(xbar = 5, ybar = 10, sx = 2, sy = 3, sxy = 0.5, n = 100,contr=TRUE)

}
