% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{regr}
\alias{regr}
\title{Regressione Lineare}
\usage{
regr(x = NULL, y = NULL, stat1 = NULL, stat2 = NULL, semp = F, ax = 2)
}
\arguments{
\item{x}{Un vettore numerico che rappresenta la variabile indipendente. Non richiesto
se `stat1` o `stat2` sono forniti.}

\item{y}{Un vettore numerico che rappresenta la variabile dipendente. Non richiesto
se `stat1` o `stat2` sono forniti.}

\item{stat1}{Un elenco contenente statistiche riassuntive dei dati. Può includere
elementi come `n`, `sumx`, `sumx2`, `sumy`, `sumy2`, e `sumxy`.}

\item{stat2}{Un'altra forma di statistiche riassuntive, simile a `stat1` ma potrebbe
includere medie e varianze direttamente.}

\item{semp}{Un booleano che indica se utilizzare una versione semplificata dei calcoli.
Utile per grandi set di dati dove la precisione al di sotto di una certa soglia
decimale non è critica.}

\item{ax}{Numero di cifre decimali per l'arrotondamento nei calcoli e nell'output.
Valore predefinito a 2.}
}
\value{
Un elenco contenente vari componenti legati alla regressione effettuata,
inclusi coefficienti, misure di adattamento, e statistiche utili per valutazioni
successive.
<<<<<<< HEAD

\code{R2} calcola e commenta R²
=======
>>>>>>> 0dee9dad6c09ddd340f4562cac51c02a2dd84b5a
}
\description{
Questa funzione esegue una regressione lineare sui dati forniti, calcolando
varie statistiche e misure di adattamento. Può operare in diversi modi a seconda
degli argomenti forniti: direttamente su vettori `x` e `y`, o utilizzando statistiche
riassuntive.
}
\examples{
<<<<<<< HEAD
require(kableExtra)
=======
>>>>>>> 0dee9dad6c09ddd340f4562cac51c02a2dd84b5a
x <- 1:10
y <- 2 + 3 * x + rnorm(10)
risultato <- regr(x, y)
print(risultato$mx)
print(risultato$my)
<<<<<<< HEAD
print(risultato$R2())
=======
>>>>>>> 0dee9dad6c09ddd340f4562cac51c02a2dd84b5a

}
