% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate.by.count}
\alias{replicate.by.count}
\title{Replicates data set based off the count variable.}
\usage{
replicate.by.count(dataset, ...)
}
\arguments{
\item{dataset}{A data.frame.}

\item{...}{column name to replicate}
}
\description{
Replicates data set based off the count variable.
}
