% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R
\name{residuo}
\alias{residuo}
\title{Stampa il Calcolo del Residuo per una Coppia (x, y) in LaTeX}
\usage{
residuo(x, y)
}
\arguments{
\item{x}{Il valore di \eqn{x_i} per il quale si desidera calcolare il residuo.}

\item{y}{Il valore di \eqn{y_i} corrispondente a \eqn{x_i}.}
}
\value{
La funzione non ritorna un valore ma stampa direttamente l'output in formato LaTeX
        nell'ambiente di chiamata.
}
\description{
Questa funzione calcola e stampa il residuo \eqn{\hat\varepsilon_i} per una coppia di valori
\(x_i\) e \(y_i\), utilizzando i coefficienti di regressione lineare \eqn{\hat\beta_0} e
\eqn{\hat\beta_1}. L'output è in formato LaTeX e mostra i passaggi del calcolo del valore previsto
\eqn{\hat y_i} e del residuo \eqn{\hat\varepsilon_i}.
}
\details{
Si assume che i coefficienti \eqn{\hat{\beta_1}} e \eqn{\hat{\beta_0}}  siano definiti
         e accessibili nello scope da cui la funzione viene chiamata. La funzione stampa l'equazione
         per calcolare \eqn{\hat y_i}e il residuo \eqn{\hat{\varepsilon_i}}.
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100,0,.1)
ls2e(regr(x,y))
# Calcolo del residuo per la coppia (x = 10, y = 20)
residuo(10, 20)

}
