% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{round_all}
\alias{round_all}
\title{Arrotonda Tutte le Variabili Numeriche in un Ambiente}
\usage{
round_all(dig = 4, exclude = NULL, env = parent.frame())
}
\arguments{
\item{dig}{Un intero che specifica il numero di cifre decimali a cui i valori numerici devono
essere arrotondati. Il valore di default è 4.}

\item{exclude}{Un vettore di caratteri che specifica i nomi delle variabili da escludere
dall'arrotondamento. Il valore di default è NULL.}

\item{env}{L'ambiente in cui le variabili numeriche devono essere arrotondate.
Il valore di default è il frame genitore da cui viene chiamata la funzione.}
}
\value{
Nessun valore di ritorno, la funzione opera modificando direttamente le variabili
        nell'ambiente specificato.
}
\description{
Questa funzione arrotonda iterativamente tutte le variabili numeriche a un numero specificato
di cifre decimali all'interno dell'ambiente fornito. Di default, utilizza l'ambiente da cui
la funzione viene chiamata.
}
\examples{
# Supponiamo di avere alcune variabili numeriche nell'ambiente globale:
x <- 3.14159
y <- 2.71828
z <- 1.61803
# Utilizziamo round_all per arrotondarle a 2 cifre decimali, escludendo 'y'
round_all(dig = 2, exclude = c("y"))

print(x) # 3.14
print(y) # 2.71828 (non arrotondato)
print(z) # 1.62

}
