% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial-functions.R
\name{run_nb}
\alias{run_nb}
\title{Load a notebook in RStudio}
\usage{
run_nb(name = NULL)
}
\arguments{
\item{name}{the name of the notebook to load}
}
\value{

}
\description{
Load a notebook in RStudio
}
\details{
The intention is to make it easier to run notebooks
}
