% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial-functions.R
\name{run_shiny}
\alias{run_shiny}
\title{Run a shiny app in your browser}
\usage{
run_shiny(name = NULL)
}
\arguments{
\item{name}{the name of the app to run.}
}
\description{
Run a shiny app in your browser
}
\details{
The intention is to make it easier to run apps.
}
