% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R
\name{S_epsilon}
\alias{S_epsilon}
\title{Calcolo della varianza di epsilon}
\usage{
S_epsilon()
}
\description{
Questa funzione calcola la varianza di epsilon (\eqn{S_\varepsilon^2}) utilizzando le formule mostrate.
}
\details{
La funzione visualizza le seguenti equazioni utilizzando la notazione LaTeX:
\deqn{\hat{\sigma_\varepsilon}^2 = (1-r^2)\hat{\sigma_Y}^2}
\deqn{(1-r^2)\times \text{vy}}
\deqn{S_\varepsilon^2 = \frac{n} {n-2} \hat{\sigma_\varepsilon}^2}
\deqn{\frac{n} {n-2} \times \text{sh2} = \text{se2}}
}
\examples{
# Esempio di utilizzo
x <- rnorm(100)
y <- x + rnorm(100,0,.1)
ls2e(regr(x,y))
S_epsilon()

}
