% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R
\name{se_beta1}
\alias{se_beta1}
\alias{se_beta0}
\title{Stampa il Calcolo della Varianza Stimata e dell'Errore Standard per \(\\hat{\\beta}_0\)}
\usage{
se_beta1(sig_eps = TRUE)

se_beta0(sig_eps = TRUE)
}
\arguments{
\item{sig_eps}{Se TRUE, calcola prima la varianza residua corretta per poi procedere al
calcolo delle varianze e degli errori standard dei coefficienti di regressione.}
}
\value{
La funzione non ritorna un valore ma stampa direttamente l'output in formato LaTeX
        nell'ambiente di chiamata.
}
\description{
Questa funzione calcola e stampa, in formato LaTeX, la varianza stimata e l'errore standard
per il coefficiente di regressione \(\\hat{\\beta}_0\) utilizzando la varianza residua.
Il calcolo include la stima dell'errore standard basato su varianze corrette.
}
\details{
Si assume che le variabili `n`, `vy`, `vx`, `mx`, `r^2`, `sh2`, `se2` e `vb0` siano definite
         e accessibili nello scope da cui la funzione viene chiamata. I calcoli mostrano come
         varianze e errori standard sono stimati per \(\\hat{\\beta}_0\).
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100, 0, .1)
ls2e(regr(x, y))  # Assume regr and ls2e are properly defined
se_beta0(TRUE)

}
