% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLinR-package.R
\docType{package}
\name{SimpLinR-package}
\alias{SimpLinR}
\alias{SimpLinR-package}
\title{SimpLinR: Simple Linear Models}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
