% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R
\name{spline_pat}
\alias{spline_pat}
\title{Spline Smussata con Bande di Confidenza}
\usage{
spline_pat(x, y, df1 = 5, df2 = 5, n = 100, a = qnorm(0.975))
}
\arguments{
\item{x}{Un vettore numerico dei valori x.}

\item{y}{Un vettore numerico dei valori y corrispondenti a `x`.}

\item{df1}{Gradi di libertà per il primo adattamento della spline smussata. Default è 5.}

\item{df2}{Gradi di libertà per l'adattamento della spline sui residui. Default è 5.}

\item{n}{Numero di punti da utilizzare per la griglia di previsione. Default è 100.}

\item{a}{Valore critico per le bande di confidenza. Default è il 97.5-esimo percentile della distribuzione normale (\code{qnorm(.975)}).}
}
\value{
La funzione non restituisce un valore ma traccia le linee della spline smussata e delle bande di confidenza sul grafico corrente.
}
\description{
Questa funzione adatta una spline smussata ai punti dati e traccia i valori
adattati insieme alle bande di confidenza.
}
\examples{

x <- seq(0,10,length=200)
y <- sin(x) + rnorm(n = 200,0,x^2/100+.1)
plot(x,y)
spline_pat(x,y,df1=10,df2=10)
}
