% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{spline_simulation}
\alias{spline_simulation}
\title{Create bias-variance tradeoff plot}
\usage{
spline_simulation(f, sigma, n, df, resamples = 100)
}
\arguments{
\item{f}{The true function}

\item{sigma}{The noise level}

\item{n}{The training sample size}

\item{df}{The degrees of freedom to use}

\item{resamples}{The number of resamples to use}
}
\value{
The plot showing the true function and the spline fits
}
\description{
Create bias-variance tradeoff plot
}
