% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.covariance}
\alias{stat.covariance}
\title{Covariance}
\usage{
stat.covariance(x, y)
}
\arguments{
\item{x}{le vecteur X}

\item{y}{le vecteur Y}
}
\value{
la covariance entre X et Y
}
\description{
Donne la covariance.\cr
 Definition 2.6 page 44
}
\details{
Notre objectif est d’obtenir une valeur numerique exprimant le degre
 de liaison qui peut exister entre deux variables X et Y.\cr\cr

Cov(X, Y) = sum((x - mean(x)) * (y - mean(y)))
}
