% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R, R/unamur.R
\name{stat.description}
\alias{stat.description}
\title{Informations statistique univarié}
\usage{
stat.description(x)

stat.description(x)
}
\arguments{
\item{x}{le vecteur.}

\item{x}{le vecteur.}
}
\description{
Permet d'avoir toutes les informations statistique univarié d'un vecteur.
}
\seealso{
\code{\link{mean}}

\code{\link{stat.var_nc}}

\code{\link{stat.var}}

\code{\link{stat.sd_nc}}

\code{\link{stat.sd}}

\code{\link{median}}

\code{\link{stat.mode}}

\code{\link{mean}}

\code{\link{stat.var_nc}}

\code{\link{stat.var}}

\code{\link{stat.sd_nc}}

\code{\link{stat.sd}}

\code{\link{median}}

\code{\link{stat.mode}}
}
