% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.linear_regression}
\alias{stat.linear_regression}
\title{Droite de régression}
\usage{
stat.linear_regression(x, y)
}
\arguments{
\item{x}{le vecteur X}

\item{y}{le vecteur Y}
}
\description{
Donne la droite de régression d'un vecteur Y en fonction d'un vecteur X.\cr
 Theoreme 2.1 page 48
}
\details{
Nous travaillons avec deux séries statistiques X et Y, mesurées sur un
échantillon de n individus. Si on suppose que le comportement théorique de Y en
fonction de X peut s’expliquer de la manière suivante\cr

La droite de régression de Y par rapport à X est la droite\cr
Y = a * X + b\cr\cr
a = Cov(X, Y) / S^2(X)\cr
b = Y - a * X
}
\examples{
Les cotes de 9 étudiants en Janvier et en fin d’année sont les suivantes.
Janvier <- c(77, 50, 71, 72, 81, 94, 96, 99, 67)
Juin <- c(82, 66, 78, 34, 47, 85, 99, 99, 68)

- Déterminez la droite de régression de Y en fonction X
stat.linear_regression(Janvier, Juin)
}
\seealso{
\code{\link{stat.covariance}}
}
