% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.linear_regression.x}
\alias{stat.linear_regression.x}
\title{Valeur de X par rapport à Y pour une droite de régression}
\usage{
stat.linear_regression.x(x, y, y_value)
}
\arguments{
\item{x}{le vecteur X}

\item{y}{le vecteur Y}

\item{y_valeur}{la valeur de Y}
}
\value{
la valeur de X
}
\description{
Donne une valeur de X par rapport à Y pour une droite de régression basé sur un
vecteur X et un vecteur Y.
}
\examples{
Les cotes de 9 étudiants en Janvier et en fin d'année sont les suivantes.
Janvier <- c(77, 50, 71, 72, 81, 94, 96, 99, 67)
Juin <- c(82, 66, 78, 34, 47, 85, 99, 99, 68)

- En utilisant le modèle de liaison entre la note de janvier et celle de
juin, estimez la cote de janvier d'un étudiant ayant eu 78 en juin
stat.linear_regression.x(Janvier, Juin, 78)
}
