% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.linear_regression.y}
\alias{stat.linear_regression.y}
\title{Valeur de Y par rapport à X pour une droite de régression}
\usage{
stat.linear_regression.y(x, y, x_value)
}
\arguments{
\item{x}{le vecteur X}

\item{y}{le vecteur Y}

\item{x_valeur}{la valeur de X}
}
\value{
la valeur de Y
}
\description{
Donne une valeur de Y par rapport à X pour une droite de régression basé sur un
vecteur X et un vecteur Y.
}
\examples{
Les cotes de 9 étudiants en Janvier et en fin d'année sont les suivantes.
Janvier <- c(77, 50, 71, 72, 81, 94, 96, 99, 67)
Juin <- c(82, 66, 78, 34, 47, 85, 99, 99, 68)

- En utilisant le modèle de liaison entre la note de janvier et celle de
juin, estimez la cote de juin d'un étudiant ayant eu 85 en janvier.
stat.linear_regression.y(Janvier, Juin, 85)
}
