% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.mode}
\alias{stat.mode}
\title{Mode}
\usage{
stat.mode(x)
}
\arguments{
\item{x}{le vecteur}
}
\description{
Donne le mode d'un vecteur.
}
\details{
Le mode de la série statistique discrète (X) = (X(i))i=1,...,n,
notée Mo(X) est la valeur xi dont la fréquence est maximale.\cr
Dès lors, on distinguera\cr
 — les distributions unimodales avec un seul mode,\cr
 — des distributions plurimodales avec plusieurs modes.\cr
}
