% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{stat}
\alias{stat}
\title{Creates Histogram, Box plot and Numerical Summary}
\usage{
stat(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Creates Histogram, Box plot and Numerical Summary
}
