% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R, R/unamur.R
\name{stat.sd}
\alias{stat.sd}
\title{Ecart-type corrigé}
\usage{
stat.sd(x)

stat.sd(x)
}
\arguments{
\item{x}{le vecteur}

\item{x}{le vecteur}
}
\description{
Donne l'écart-type empirique corrigé d'un vecteur.\cr
 Equation 1.28 page 22
}
