% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.sd_nc}
\alias{stat.sd_nc}
\title{Ecart-type non corrigé}
\usage{
stat.sd_nc(x)
}
\arguments{
\item{x}{le vecteur}
}
\description{
Donne l'écart-type empirique non corrigé d'un vecteur.\cr
 Equation 1.27 page 22
}
