% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.var_nc}
\alias{stat.var_nc}
\title{Variance non corrigé}
\usage{
stat.var_nc(x)
}
\arguments{
\item{x}{le vecteur}
}
\description{
Donne la variance empirique non corrigé d'un vecteur.\cr
 Equation 1.25 page 22
}
