% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_base.R
\name{stat_base}
\alias{stat_base}
\title{Stat Base - Analisi Statistica di Base}
\usage{
stat_base(samp, brk)
}
\arguments{
\item{samp}{Un vettore di dati numerici campionari generati dalla funzione `genera_dati`.}

\item{brk}{Un vettore numerico che specifica gli intervalli per la categorizzazione dei dati campionari.}
}
\value{
Un elenco contenente le seguenti componenti:
  \itemize{
    \item \code{dat2}: Un data frame con statistiche di base per ogni intervallo.
    \item \code{dat3}: Un data frame simile a \code{dat2} ma include anche i totali di colonna.
    \item \code{F_print(x, verso, x2)}: calcola l'area tra \code{x} e \code{x2}.
    \item \code{histp(axes = F)}: genera l'istogramma delle densità percentuali.
    \item \code{h.int(x1, x2, density = 20, ...)}: aggiunge l'area tratteggiata tra \code{x1} e \code{x2}.
    \item \code{percentile(p)}: calcola il percentile \code{p}.
    \item Altre funzioni per analisi statistiche dettagliate (\code{F.int}, \code{H.int}, \code{Q.int}).
  }
}
\description{
Questa funzione esegue un'analisi statistica di base sui dati campionari forniti e ritorna
un insieme di funzioni e statistiche dettagliate. La funzione `F_print` inclusa nell'output
permette di stampare risultati specifici dell'analisi.
}
\examples{
brk <- c(0, 10, 20, 30, 40)
hhh <- c( 2, 4,  2,   1)
samp <- genera_dati(brk, hhh=hhh, n=100)
risultati <- stat_base(samp, brk)
risultati$F_print(x=15, verso="<")
risultati$histp(TRUE)
risultati$h.int(5,27,density=20)
risultati$percentile(.75)
}
