\name{StatComp20008-package}
\alias{StatComp20008-package}
\alias{StatComp20008}
\docType{package}
\title{
  Carrying out Count Five Test for equal Variance.
}
\description{
  The two sample "Count Five" test for equality of variance is based on the maximum number of extreme points.
}
\details{
  The funtion CountFive(x,y)is applied to vector x and y when sample size is equal.While hat_alpha(x,y,m=1e3) estimate the significance level of the CountFive test.
}
\author{
Yangshuhua.

Maintainer: Yangshuhua <shuashua0608@mail.ustc.edu.cn>
}
\references{
 Statistical Computing with R.
}
\keyword{ package } 

