% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.R
\name{symmetric.contingency.table.to.dataset}
\alias{symmetric.contingency.table.to.dataset}
\title{Creates a data set from a symmetric contingency table}
\usage{
symmetric.contingency.table.to.dataset(table, col.names, num.way.table = 2)
}
\arguments{
\item{table}{A contingency table}

\item{col.names}{The dataset columns}

\item{num.way.table}{optional:}
}
\value{
dataset from contingency table
}
\description{
Creates a data set from a symmetric contingency table
}
\examples{

}
