% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions-unimore.R, R/main-functions.R
\name{tabl}
\alias{tabl}
\title{Crea una Tabella Stilizzata con kable e kableExtra}
\usage{
tabl(x, ...)

tabl(x, ...)
}
\arguments{
\item{x}{Un oggetto data frame o matrice da visualizzare come tabella.}

\item{...}{Parametri addizionali passati a `kable`.}
}
\value{
Restituisce un oggetto di tipo tabella stilizzata pronto per essere stampato
o visualizzato.

Restituisce un oggetto di tipo tabella stilizzata pronto per essere stampato
o visualizzato.
}
\description{
Questa funzione genera una tabella stilizzata partendo da un data frame o una matrice,
utilizzando `kable` di **knitr** per la creazione della tabella e `kable_styling` di
**kableExtra** per l'applicazione dello stile.

Questa funzione genera una tabella stilizzata partendo da un data frame o una matrice,
utilizzando `kable` di **knitr** per la creazione della tabella e `kable_styling` di
**kableExtra** per l'applicazione dello stile.
}
\details{
La funzione imposta `booktabs = TRUE` per utilizzare la formattazione 
Booktabs, che migliora l'aspetto delle tabelle LaTeX e HTML, e `escape = FALSE` 
per evitare l'escape automatico di caratteri speciali LaTeX. Inoltre, applica
specifiche opzioni di styling attraverso `kable_styling` per controllare l'aspetto
finale della tabella, come `full_width = FALSE` e l'opzione `latex_options` impostata
su "HOLD_position" per gestire la posizione della tabella.

La funzione imposta `booktabs = TRUE` per utilizzare la formattazione 
Booktabs, che migliora l'aspetto delle tabelle LaTeX e HTML, e `escape = FALSE` 
per evitare l'escape automatico di caratteri speciali LaTeX. Inoltre, applica
specifiche opzioni di styling attraverso `kable_styling` per controllare l'aspetto
finale della tabella, come `full_width = FALSE` e l'opzione `latex_options` impostata
su "HOLD_position" per gestire la posizione della tabella.
}
\examples{
# Assicurati di avere i pacchetti knitr e kableExtra installati e caricati
if (require(knitr) && require(kableExtra)) {
  data(mtcars)
  tabl(mtcars[1:5, 1:5])
}

# Assicurati di avere i pacchetti knitr e kableExtra installati e caricati
if (require(knitr) && require(kableExtra)) {
  data(mtcars)
  tabl(mtcars[1:5, 1:5])
}

}
