% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R, R/main-functions-unimore.R
\name{ztest_pi}
\alias{ztest_pi}
\alias{ztest_mu}
\alias{ttest_mu}
\title{Test Z e Test T: Proporzione e Media}
\usage{
ztest_pi(sn, n, p0, h1 = "\\\\neq", alpha = 0.05)

ztest_mu(muh, s, n, mu0, h1 = "\\\\neq", alpha = 0.05, um = "", pvalue = T)

ttest_mu(muh, sh, n, mu0, h1 = "\\\\neq", alpha = 0.05, um = "")

ztest_pi(sn, n, p0, h1 = "\\\\neq", alpha = 0.05)

ztest_mu(muh, s, n, mu0, h1 = "\\\\neq", alpha = 0.05, um = "", pvalue = T)

ttest_mu(muh, sh, n, mu0, h1 = "\\\\neq", alpha = 0.05, um = "")
}
\arguments{
\item{sn}{Numero di successi o media campionaria.}

\item{n}{Dimensione del campione.}

\item{p0}{Proporzione attesa nella popolazione per `ztest_pi`.}

\item{h1}{Ipotesi alternativa: può essere "\\neq", ">", "<".}

\item{alpha}{Livello di significatività.}

\item{muh}{Media campionaria per `ztest_mu` e `ttest_mu`.}

\item{s}{Deviazione standard della popolazione per `ztest_mu`, o deviazione standard campionaria per `ttest_mu`.}

\item{mu0}{Media attesa nella popolazione per `ztest_mu` e `ttest_mu`.}

\item{um}{Unità di misura (opzionale) per migliorare la leggibilità dell'output.}

\item{pvalue}{Se `TRUE`, mostra il p-value del test (solo per `ztest_mu`).}
}
\description{
Fornisce metodi per eseguire test z su una proporzione (`ztest_pi`), test z su una media con varianza nota (`ztest_mu`), e test t su una media con varianza incognita (`ttest_mu`).

Fornisce metodi per eseguire test z su una proporzione (`ztest_pi`), test z su una media con varianza nota (`ztest_mu`), e test t su una media con varianza incognita (`ttest_mu`).
}
\details{
\itemize{
\item `ztest_pi` utilizza il test z per valutare se la proporzione osservata in un campione differisce significativamente dalla proporzione attesa, sotto l'ipotesi nulla.
\item `ztest_mu` applica il test z per determinare se la media campionaria differisce significativamente dalla media attesa, assumendo che la varianza della popolazione sia nota.
\item `ttest_mu` impiega il test t per testare differenze nella media campionaria rispetto a una media attesa, quando la varianza della popolazione non è nota.
}

\itemize{
\item `ztest_pi` utilizza il test z per valutare se la proporzione osservata in un campione differisce significativamente dalla proporzione attesa, sotto l'ipotesi nulla.
\item `ztest_mu` applica il test z per determinare se la media campionaria differisce significativamente dalla media attesa, assumendo che la varianza della popolazione sia nota.
\item `ttest_mu` impiega il test t per testare differenze nella media campionaria rispetto a una media attesa, quando la varianza della popolazione non è nota.
}
}
\examples{
# Test Z per una proporzione
ztest_pi(sn = 55, n = 100, p0 = 0.5, h1 = "\\\\neq", alpha = 0.05)

# Test Z per una media con varianza nota
ztest_mu(muh = 5, s = 1.5, n = 30, mu0 = 5, h1 = "\\\\neq", alpha = 0.05)

# Test T per una media con varianza incognita
ttest_mu(muh = 5, sh = 1.5, n = 30, mu0 = 5, h1 = "\\\\neq", alpha = 0.05)

# Test Z per una proporzione
ztest_pi(sn = 55, n = 100, p0 = 0.5, h1 = "\\\\neq", alpha = 0.05)

# Test Z per una media con varianza nota
ztest_mu(muh = 5, s = 1.5, n = 30, mu0 = 5, h1 = "\\\\neq", alpha = 0.05)

# Test T per una media con varianza incognita
ttest_mu(muh = 5, sh = 1.5, n = 30, mu0 = 5, h1 = "\\\\neq", alpha = 0.05)

}
