% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{tlc}
\alias{tlc}
\title{Teorema del Limite Centrale (TLC): Applicazioni}
\usage{
tlc(tipo, x1, x2 = NULL, verso, mu = F, s2 = NULL, n)
}
\arguments{
\item{tipo}{Specifica il tipo di calcolo: "somma", "media", "prop" (proporzione), "pois_media" (media di Poisson), o "pois_somma" (somma di Poisson).}

\item{x1, x2}{Valori di interesse per il calcolo della probabilità. Se `x2` è fornito, il calcolo viene eseguito su un intervallo.}

\item{verso}{Specifica la direzione della probabilità (\eqn{\leq}, \eqn{\geq}).}

\item{mu}{Valore atteso \eqn{\mu} delle variabili casuali, o \eqn{\pi} per la proporzione, o \eqn{\lambda} per Poisson.}

\item{s2}{Varianza \eqn{\sigma^2} delle variabili casuali (non necessario per "prop").}

\item{n}{Numero di variabili casuali indipendenti e identicamente distribuite (IID).}
}
\description{
La funzione `tlc` applica il Teorema del Limite Centrale a diverse situazioni: somma di variabili casuali, calcolo della media, proporzione, e la somma o media di variabili casuali di Poisson.
}
\details{
La funzione fornisce un'interfaccia unificata per dimostrare come il TLC può essere applicato in diversi contesti statistici, generando l'approssimazione normale di diverse distribuzioni di probabilità basate su condizioni specificate.
}
\examples{
# Somma di variabili casuali con varianza nota
tlc(tipo = "somma", x1 = 5, verso = "<", mu = 2, s2 = 1, n = 30)

# Calcolo della media di variabili casuali
tlc(tipo = "media", x1 = 2.5, verso = "<", mu = 2, s2 = 1, n = 50)

# Proporzione in una distribuzione binomiale approssimata
tlc(tipo = "prop", x1 = 0.6, verso = ">", mu = 0.5, n = 100)

# Media di variabili casuali di Poisson
tlc(tipo = "pois_media", x1 = 3, verso = "<", mu = 2, n = 50)

# Somma di variabili casuali di Poisson
tlc(tipo = "pois_somma", x1 = 150, verso = ">", mu = 3, n = 50)

}
