% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R
\name{TSS}
\alias{TSS}
\title{Stampa i Componenti della Varianza Totale dei Dati in LaTeX}
\usage{
TSS()
}
\value{
La funzione non ritorna un valore ma stampa direttamente l'output in formato LaTeX
        nell'ambiente di chiamata.
}
\description{
Questa funzione calcola e stampa il Total Sum of Squares (TSS), Explained Sum of Squares (ESS),
e Residual Sum of Squares (RSS) utilizzando il coefficiente di determinazione \(R^2\). La funzione
stampa queste statistiche in formato LaTeX, illustrando come TSS è suddiviso in ESS e RSS.
}
\details{
Si assume che le variabili `n`, `vy`, e `r^2` (numero di osservazioni, varianza di Y,
         e il quadrato del coefficiente di correlazione, rispettivamente) siano definite e
         accessibili nello scope da cui la funzione viene chiamata. La funzione dimostra la
         relazione \(TSS = ESS + RSS\).
}
\examples{
# Calcolo e stampa di TSS, ESS e RSS supponendo che n, vy e r^2 siano già definiti
TSS()

}
