% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R
\name{ttest_beta}
\alias{ttest_beta}
\title{Conduce e Stampa l'Analisi di un t-Test per un Coefficiente di Regressione in LaTeX}
\usage{
ttest_beta(
  cof,
  bj0,
  h1 = "\\\\neq",
  alpha = c(1/10, 5/100, 1/100, 1/1000),
  SE = F,
  pv_only = F
)
}
\arguments{
\item{cof}{Indice del coefficiente su cui eseguire il test (0 per \eqn{\hat{\beta}_0}, 1 per  \eqn{\hat{\beta}_1}).}

\item{bj0}{Valore ipotizzato per il coefficiente sotto l'ipotesi nulla.}

\item{h1}{L'ipotesi alternativa del test (\(\\neq\), \(>\), o \(<\)).}

\item{alpha}{Un vettore di livelli di significatività da utilizzare per il test.}

\item{SE}{Se TRUE, esegue e stampa il calcolo dell'errore standard del coefficiente specificato.}
}
\value{
La funzione non ritorna un valore ma stampa direttamente l'output in formato LaTeX
        e produce grafici e p-value associati al test, nell'ambiente di chiamata.
}
\description{
Questa funzione esegue un t-test su uno dei coefficienti di regressione \eqn{\hat{\beta}_0} o \eqn{\hat\beta_1},
stampando il calcolo della statistica del test e altri output correlati in formato LaTeX. 
La funzione può anche eseguire e stampare il calcolo dell'errore standard del coefficiente se specificato.
}
\details{
Si assume che le variabili `b0`, `b1`, `vb0`, `vb1`, e `n` siano definite e accessibili
         nello scope da cui la funzione viene chiamata. Include funzioni di supporto come `A_`, `C_`,
         `graf`, e `p_value` che dovrebbero essere definite per eseguire analisi aggiuntive e stampa.
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100, 0, .1)
ls2e(regr(x, y))  # Assume regr and ls2e are properly defined
ttest_beta(cof = 0, bj0 = 0)  # Test su beta_0
ttest_beta(cof = 1, bj0 = 1, SE = TRUE)  # Test su beta_1 con errore standard

}
