% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{visualize_knn}
\alias{visualize_knn}
\title{Weighted K-Nearest Neighbors visualization}
\usage{
visualize_knn(
  formula,
  training_data,
  test_data,
  k = 1,
  weights = NULL,
  algorithm = c("kd_tree", "cover_tree", "brute")
)
}
\arguments{
\item{formula}{Formula like one would use for \code{lm()}}

\item{training_data}{The training data, in the form of a data frame}

\item{test_data}{The test data, in the form of a data frame}

\item{k}{The number of nearest neighbors}

\item{weights}{A vector of weights}

\item{algorithm}{Nearest neighbor search algorithm.}
}
\value{
A \code{ggplot} object
}
\description{
Carry out weighted KNN and plot the results.
}
