% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-quadro.R
\name{chi_print}
\alias{chi_print}
\alias{chi_print_conf}
\alias{chi_test}
\alias{chi_conf}
\title{Test e Visualizzazione del Chi Quadrato}
\usage{
chi_print(dat, nome_x, nome_y, print = T)

chi_print_conf(Freq_c, Freq_0, X, Y)

chi_test(dat, alpha)

chi_conf(Freq_c, Freq_0, X, Y, alpha = 0.05)
}
\arguments{
\item{dat}{Matrice o data frame contenente i dati per `chi_print` e `chi_test`.}

\item{nome_x, }{nome_y Nomi delle variabili o delle categorie per `chi_print`.}

\item{print}{Se `TRUE`, stampa la tabella di contingenza per `chi_print`.}

\item{Freq_c, }{Freq_0 Frequenze osservate e attese per `chi_print_conf` e `chi_conf`.}

\item{X, }{Y Nomi delle variabili o delle categorie per `chi_print_conf` e `chi_conf`.}

\item{alpha}{Livello di significatività per `chi_test` e `chi_conf`.}

\item{mu1, }{mu2 Medie campionarie per i test su due campioni.}

\item{s1h, }{s2h Deviazioni standard campionarie per i test su due campioni.}

\item{n1, }{n2 Dimensioni dei campioni per i test su due campioni.}

\item{h1}{Ipotesi alternativa: può essere "\\neq", ">", "<".}

\item{et}{Specifica se il test assume varianze eterogenee (`TRUE`) o omogenee (`FALSE`).}

\item{a, }{b Etichette per i gruppi, utili per l'output di `test_2c`.}

\item{um}{Unità di misura (opzionale), utile per migliorare la leggibilità dell'output di `test_2c`.}
}
\description{
Fornisce metodi per visualizzare tabelle di contingenza e per eseguire il test del chi quadrato per indipendenza e conformità.
}
\details{
Le funzioni fornite includono:
\itemize{
  \item \code{chi_print}: Visualizza una tabella di contingenza arricchita con i totali per righe e colonne.
  \item \code{chi_print_conf}: Visualizza una tabella di confronto tra frequenze osservate e attese, utile per l'analisi di conformità.
  \item \code{chi_test}: Esegue il test del chi quadrato per indipendenza su una tabella di contingenza, valutando l'associazione tra due variabili categoriche.
  \item \code{chi_conf}: Esegue il test del chi quadrato per conformità, confrontando le frequenze osservate con quelle attese basate su una distribuzione specificata.
}

Queste funzioni sono progettate per facilitare l'analisi statistica di dati categorici, permettendo agli utenti di eseguire test di ipotesi e visualizzare i risultati in modo intuitivo.
}
\examples{

dat <- matrix(c(15,5,12,10,35,15),2,byrow = T)

nome_x<- c("Rai","Mediaset","La7")
nome_y <- c("Laureato","Non Laureato")

dat          <- chi_print(dat,nome_x = nome_x,nome_y = nome_y)[[1]]
dat_print    <- chi_print(dat,nome_x = nome_x,nome_y = nome_y)[[2]]

tabl(dat_print)

cat(chi_test(dat = dat,alpha = 0.01))

fx <- sample
Freqc <- sample(1:25,6,T)
Freq0 <- (numeric(6)+1/6)*100
chi_print_conf(Freqc,Freq0,X = letters[1:6],Y = letters[1:2])
chi_conf(Freqc,Freq0,X = letters[1:6],Y = letters[1:2])
}
