% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferenza.R, R/test.R
\name{idc}
\alias{idc}
\title{Calcolo dell'Intervallo di Confidenza}
\usage{
idc(xm, sd = NULL, alpha = 0.05, n, dist_, mus = NULL, ss = NULL)

idc(xm, sd = NULL, alpha = 0.05, n, dist_, mus = NULL, ss = NULL)
}
\arguments{
\item{xm}{Media campionaria o somma delle successi.}

\item{sd}{Deviazione standard della popolazione, se nota.}

\item{alpha}{Livello di significatività per l'intervallo di confidenza.}

\item{n}{Dimensione del campione.}

\item{dist_}{Tipo di distribuzione ("z" per Normale, "t" per t di Student).}

\item{mus}{Media della popolazione, se nota (utilizzato per calcoli specifici).}

\item{ss}{Varianza campionaria, se disponibile.}
}
\description{
Calcola l'intervallo di confidenza per la media o la proporzione di una popolazione, utilizzando la distribuzione Normale (z) o t di Student.

Calcola l'intervallo di confidenza per la media o la proporzione di una popolazione, utilizzando la distribuzione Normale (z) o t di Student.
}
\details{
La funzione `idc` determina l'intervallo di confidenza utilizzando la formula appropriata in base alla distribuzione specificata.
Per le proporzioni, la deviazione standard viene calcolata internamente se non fornita. 
Per la distribuzione t, la deviazione standard campionaria viene corretta per la dimensione del campione.

La funzione `idc` determina l'intervallo di confidenza utilizzando la formula appropriata in base alla distribuzione specificata.
Per le proporzioni, la deviazione standard viene calcolata internamente se non fornita. 
Per la distribuzione t, la deviazione standard campionaria viene corretta per la dimensione del campione.
}
\examples{
# Intervallo di confidenza per la media con distribuzione Normale
idc(xm = 5, sd = 2, alpha = 0.05, n = 30, dist_ = "z")

# Intervallo di confidenza per la media con distribuzione t
idc(xm = 5, ss = 4, alpha = 0.05, n = 30, dist_ = "t")

# Intervallo di confidenza per la proporzione
idc(xm = 120, alpha = 0.05, n = 200, dist_ = "z")

# Intervallo di confidenza per la media con distribuzione Normale
idc(xm = 5, sd = 2, alpha = 0.05, n = 30, dist_ = "z")

# Intervallo di confidenza per la media con distribuzione t
idc(xm = 5, ss = 4, alpha = 0.05, n = 30, dist_ = "t")

# Intervallo di confidenza per la proporzione
idc(xm = 120, alpha = 0.05, n = 200, dist_ = "z")

}
