% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotowce.R
\name{kod_perm_pary}
\alias{kod_perm_pary}
\title{Szkielet: Permutacje dla PAR (Zależne - Wersja z dwiema pętlami)}
\usage{
kod_perm_pary()
}
\description{
Test permutacyjny dla par (np. klasyczne vs nowe).
Wersja zagnieżdżona (for w for), dokładnie jak na screenach z zajęć.
}
