% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilita.R
\name{p_aub}
\alias{p_aub}
\title{Calcolo della probabilità dell'unione di due eventi}
\usage{
p_aub(pa, pb, A = "A", B = "B", dig = 4)
}
\arguments{
\item{pa}{La probabilità dell'evento \(A\).}

\item{pb}{La probabilità dell'evento \(B\).}

\item{A}{Il nome dell'evento \(A\) (opzionale, predefinito è "A").}

\item{B}{Il nome dell'evento \(B\) (opzionale, predefinito è "B").}

\item{dig}{Il numero di cifre decimali per l'arrotondamento (opzionale, predefinito è 4).}
}
\description{
Questa funzione calcola la probabilità dell'unione di due eventi \(A\) e \(B\) dati le probabilità individuali \(P(A)\) e \(P(B)\).
}
\details{
La probabilità dell'unione di due eventi \(A\) e \(B\) è data dalla formula:
\deqn{P(A \cup B) = P(A) + P(B) - P(A \cap B)}
dove \(P(A \cap B) = P(A) \cdot P(B)\) se \(A\) e \(B\) sono indipendenti.

La funzione stampa la formula e il risultato calcolato in notazione LaTeX.
}
\examples{
p_aub(0.5, 0.3)
p_aub(0.5, 0.3, A = "Evento1", B = "Evento2", dig = 2)

}
