% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{round_all}
\alias{round_all}
\title{Arrotonda Tutte le Variabili Numeriche in un Ambiente}
\usage{
round_all(dig = 4, env = parent.frame())
}
\arguments{
\item{dig}{Un intero che specifica il numero di cifre decimali a cui i valori numerici devono
essere arrotondati. Il valore di default è 4.}

\item{env}{L'ambiente in cui le variabili numeriche devono essere arrotondate.
Il valore di default è il frame genitore da cui viene chiamata la funzione.}
}
\value{
Nessun valore di ritorno, la funzione opera modificando direttamente le variabili
        nell'ambiente specificato.
}
\description{
Questa funzione arrotonda iterativamente tutte le variabili numeriche a un numero specificato
di cifre decimali all'interno dell'ambiente fornito. Di default, utilizza l'ambiente da cui
la funzione viene chiamata.
}
\examples{
# Supponiamo di avere alcune variabili numeriche nell'ambiente globale:
x <- 3.14159
y <- 2.71828
# Utilizziamo round_all per arrotondarle a 2 cifre decimali
round_all(dig = 2)

}
