% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT.R
\name{STAT}
\alias{STAT}
\title{Launch 'STAT' Interface}
\usage{
STAT()
}
\value{
Nothing
}
\description{
STAT() loads interactive user interface built using R 'shiny'.
}
\details{
Start STAT

The interactive user interface is to provide an easy way for basic statistical analysis and downloading plots.
}
\examples{
if(interactive()){
library(rmarkdown)
STAT()
}
}
\keyword{STAT}
